application creation available via the dashboard
New applications are created on the dashboard using the brizo api. There isn't any validation within this commit, but UUIDs are automatically generated on the brizo side via the api. There is also a need for page refreshes following a form submit.

This commit and its parent should probably be squashed and cleaned since this is semi wip.