bootstrapped service created alongside with env
Newly created environments will have a service bootstrapped to work with them.
Added a product-mocker service to be running with tests as part of bootstrapping the brizo namespace.

There are some issues with how we handle getting application info before persisting the environment, so we'll need to come up with a better way to handle this.